uploads to s3 are now done through this client app :party:
